Rework resource loading
