--- include/resource/assimp_init.h
@@ -1,17 +1,0 @@
-#include <assimp/LogStream.hpp>
-
-namespace arch {
-
-class SpdlogStream : public Assimp::LogStream {
-public:
-    void write(const char *msg) override;
-};
-
-class AssimpInitializer {
-public:
-    static void init();
-private:
-    static void init_logger();
-};
-
-}
