--- include/resource/assimp_texture_loader.h
@@ -1,10 +1,0 @@
-#include "resource/texture_loader.h"
-
-namespace arch {
-
-class AssimpTextureLoader : public TextureLoader {
-public:
-    Texture read_file(const std::filesystem::path &path) const override;
-};
-
-}
