+++ include/resource/resource_loader.h
@@ -1,0 +1,7 @@
+#include <filesystem>
+
+template<typename Resource>
+class ResourceLoader {
+public:
+    virtual Resource read_file(const std::filesystem::path &filepath);
+};
